refactor(deps): better separation for algorithm
